test(python): isolate v6 job type assertions
